portfolio-fastapi/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
influxdb-client>=1.39.0
pandas>=1.5.0
numpy>=1.20.0
python-dotenv>=1.0.0
pyarrow>=10.0.0
pytz>=2023.3
sqlalchemy>=1.4.0 # Add SQLAlchemy
psycopg2-binary>=2.9.0 # Add PostgreSQL driver